User feedback page

+++ grails-app/domain/appwebtfg/UserFeedback.groovy
@@ -3,17 +3,17 @@ package appwebtfg
 class UserFeedback {
 
     static belongsTo = [user : User]
-    int clicksSearch
-    int clicksRecommend
-    int clicksRelatedToQuery
-    int clicksRelatedToCourse
-    int addToListSearch
-    int addToListRecommend
-    int addToListRelatedToQuery
-    int addToListRelatedToCourse
-    int notInterestedCoursera
-    int notInterestedUdemy
-    int notInterestedUdacity
+    long clicksSearch
+    long clicksRecommend
+    long clicksRelatedToQuery
+    long clicksRelatedToCourse
+    long addToListSearch
+    long addToListRecommend
+    long addToListRelatedToQuery
+    long addToListRelatedToCourse
+    long notInterestedCoursera
+    long notInterestedUdemy
+    long notInterestedUdacity
 
     static constraints = {
     }
